% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_feature_engineering.R
\name{add_new_feature}
\alias{add_new_feature}
\title{Add Magnitude Type and Storm Duration Columns}
\usage{
add_new_feature(data)
}
\arguments{
\item{data}{A data frame containing the storm dataset.}
}
\value{
The input data frame with the new columns added and rows with NA 'magnitude_type' dropped.
}
\description{
This function adds categorical and numerical columns 'magnitude_type' and 'storm_duration'
to the storm dataset based on the values in the 'magnitude' column and the difference between
'event_begin_time' and 'event_end_time'. It also drops rows where 'magnitude_type' is NA.
}
\examples{
\dontrun{
storm_data <- load_and_clean_storm_data()
storm_data <- add_new_feature(storm_data)
head(storm_data)
}

}
