% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualize.R
\name{display_column_summary}
\alias{display_column_summary}
\title{Display Summary of DataFrame Columns}
\usage{
display_column_summary(df)
}
\arguments{
\item{df}{The dataframe for which to display the summary.}
}
\value{
None
}
\description{
This function displays a summary of the columns and their data types in the
provided dataframe.
}
\examples{
\dontrun{
display_column_summary(cleaned_storm_data)
}
}
