% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_and_clean_storm_data}
\alias{load_and_clean_storm_data}
\title{Load and Clean Storm Dataset from Package}
\usage{
load_and_clean_storm_data(file_name)
}
\arguments{
\item{file_name}{A character string specifying the name of the CSV file (including extension)
located in the 'inst' folder of the package.}
}
\value{
A cleaned and transformed data frame of the storm dataset.
}
\description{
This function demonstrates how to load and clean a storm dataset from a CSV file
located in the 'inst' folder of the package.
}
\examples{
\dontrun{
cleaned_storm_data <- load_and_clean_storm_data(file_name = "storm_2013.csv")
head(cleaned_storm_data)
}

}
