% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_csv_from_package}
\alias{load_csv_from_package}
\title{Load CSV Data from Package}
\usage{
load_csv_from_package(file_name)
}
\arguments{
\item{file_name}{A character string specifying the name of the CSV file (including extension)}
}
\value{
A data frame containing the contents of the loaded CSV file.
}
\description{
This function loads data in CSV format from the 'inst' folder of the package.
}
\examples{
\dontrun{
loaded_data <- load_csv_from_package(file_name = "data_file.csv")
head(loaded_data)
}

}
