% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_cleaned_datasets}
\alias{merge_cleaned_datasets}
\title{Merge Cleaned Datasets}
\usage{
merge_cleaned_datasets(df1, df2)
}
\arguments{
\item{df1}{The first cleaned dataframe.}

\item{df2}{The second cleaned dataframe.}
}
\value{
A merged dataframe.
}
\description{
This function merges two cleaned dataframes.
}
\examples{
\dontrun{
merged_data <- merge_cleaned_datasets(cleaned_data1, cleaned_data2)
head(merged_data)
}

}
